more cleanup * move BufferedProcess into process library * move promptUser into notification library
